#572 added build system, fixed some broken tests
