+++ gulpfile.babel.js
@@ -1,0 +1,3 @@
+import gulp from 'gulp';
+
+import './gulp/build';
