+++ package.json
@@ -1,11 +1,13 @@
 {
   "name": "backbone-relational",
-  "main": "backbone-relational.js",
+  "main": "dist/backbone-relational.js",
   "description": "Get and set relations (one-to-one, one-to-many, many-to-one) for Backbone models",
   "homepage": "http://backbonerelational.org",
   "keywords": [
     "backbone",
     "relation",
+    "many-to-one",
+    "one-to-many",
     "nested",
     "model"
   ],
@@ -14,18 +16,33 @@
     "url": "https://github.com/PaulUithol/Backbone-relational.git"
   },
   "scripts": {
-    "test": "karma start --single-run --browsers PhantomJS"
+    "test": "node_modules/.bin/karma start --single-run --browsers PhantomJS",
+    "build": "node_modules/.bin/gulp build"
   },
   "author": "Paul Uithol <https://github.com/PaulUithol>",
   "contributors": "Listed at <https://github.com/PaulUithol/Backbone-relational/contributors>",
   "peerDependencies": {
-    "backbone": "1.2.3"
+    "backbone": "1.2.3",
+    "underscore": "1.8 - 1.8.3"
   },
   "license": "MIT",
   "version": "0.10.0",
   "devDependencies": {
+    "babel-core": "^6.8.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-es2015-rollup": "^1.1.1",
+    "babel-register": "^6.8.0",
+    "babelify": "^7.3.0",
     "backbone": "1.2.3",
     "browserify": "^13.0.0",
+    "gulp-file": "^0.3.0",
+    "gulp-filter": "^4.0.0",
+    "gulp-plumber": "^1.1.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-rollup": "^1.9.0",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-uglify": "^1.5.3",
+    "jasmine-core": "^2.4.1",
     "jquery": "^2.2.1",
     "karma": "^0.13.22",
     "karma-browserify": "^5.0.2",
@@ -33,8 +50,11 @@
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-qunit": "^0.1.9",
     "phantomjs-prebuilt": "^2.1.5",
-    "qunitjs": "^1.22.0",
-    "underscore": ">=1.7.0",
-    "watchify": "^3.7.0"
+    "rollup": "^0.26.3",
+    "rollup-plugin-babel": "^2.4.0",
+    "rollup-plugin-json": "^2.0.0"
+  },
+  "dependencies": {
+    "babel-polyfill": "^6.9.1"
   }
 }
